requirements.txt
annotated-types==0.7.0
APScheduler==3.11.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
schedule==1.2.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
